Update patch set 1
Patch Set 1:

I put this completely in the device folder because I'm not sure if the voltage thresholds are the same on all the devices (I assume they aren't, but that could be a completely off-base assumption). It can be moved to s4-common if all devices support it, and use the same voltages. I think there still needs to be some tweaking to the voltages, but I tried once and failed miserably. Any help is appreciated.

Patch-set: 1
Reviewer: Captain Throwback <6558@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Verified=0