Update patch set 2
Patch Set 2:

I think we should re-consider the -2 on this. HTC seems to be ok with re-distribution the RUU, seeing as they let it loose via twitter ( https://twitter.com/mfichter/statuses/435959285730656256 ). Seems like a convenient way to force people to update all the firmware on their phone too.

Patch-set: 2