Update patch set 7
Patch Set 7:

Is there a specific reason we can no longer mount the internal storage as mass storage since the partition change?  I'm probably in a minority, but I generally don't use external storage. It would be nice to have that option back if this was a design decision rather than technical.

Patch-set: 7